[SWM-228] fix : 게시물성향, 유저성향 필드 변경2

+++ src/main/java/com/example/tripmingle/dto/etc/UpdateBoardDTO.java
@@ -14,11 +14,11 @@ public class UpdateBoardDTO {
 	private LocalDate endDate;
 	private String language;
 	private int maxCount;
-	private double preferGender; // 선호 성별
-	private double preferSmoking; // 선호 흡연타입
-	private double preferBudget; // 선호 활동 - 예산
-	private double preferPhoto; // 선호 활동 - 사진
-	private double preferDrink; // 선호 활동 - 음주
+	private int preferGender; // 선호 성별
+	private int preferSmoking; // 선호 흡연타입
+	private int preferBudget; // 선호 활동 - 예산
+	private int preferPhoto; // 선호 활동 - 사진
+	private int preferDrink; // 선호 활동 - 음주
 	private String tripType;
 	private String title;
 	private String content;
